backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
httpx==0.27.2
python-multipart==0.0.9
numpy>=1.24.0
python-dotenv>=1.0.0
deep-translator>=1.11.4
# ML and XAI dependencies
scikit-learn>=1.3.0
lightgbm>=4.1.0
pandas>=2.0.0
shap>=0.43.0
lime>=0.2.0.1
joblib>=1.3.0